\name{h2oEnsemble-package}
\alias{h2oEnsemble-package}
\alias{h2oEnsemble}
\docType{package}
\title{
H2O Ensemble Package
}
\description{
This package implements the super learner ensemble (stacking) algorithm using the H2O R interface to provide base leanrning algorithms. 
}
\details{
\tabular{ll}{
Package: \tab h2oEnsemble\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2014-10-08\cr
License: \tab Apache License (== 2.0)\cr
}
}
\author{
Erin LeDell

Maintainer: Erin LeDell <ledell@berkeley.edu>
}
\references{
van der Laan, M. J., Polley, E. C. and Hubbard, A. E. (2007) Super Learner, Statistical Applications of Genetics and Molecular Biology, 6, article 25. \cr
\url{http://dx.doi.org/10.2202/1544-6115.1309}
}
\keyword{ models }
\seealso{
\code{\link[SuperLearner:SuperLearner]{SuperLearner}}, \code{\link[subsemble:subsemble]{subsemble}}
}
\examples{
See \code{\link[h2oEnsemble:h2o.ensemble]{h2o.ensemble}} for examples.
}
