\name{print.h2o.ensemble}
\alias{print.h2o.ensemble}
\title{
  Print method for an 'h2o.ensemble' object
}
\description{
  Print metadata for an \code{\link[h2oEnsemble]{h2o.ensemble}} fit.  
}
\usage{
  \method{print}{h2o.ensemble}(x, ...)
}
\arguments{
  \item{x}{
    An object of class 'h2o.ensemble', which is returned from the \code{\link[h2oEnsemble]{h2o.ensemble}} function.
  }
  \item{...}{
    Additional arguments passed on to the function.
  }
}

\author{
  Erin LeDell \email{erin@h2o.ai}
}

\seealso{
  \code{\link[h2oEnsemble:h2o.ensemble]{h2o.ensemble}}
}

