\name{h2oEnsemble-package}
\alias{h2oEnsemble-package}
\alias{h2oEnsemble}
\docType{package}
\title{
H2O Ensemble Package
}
\description{
H2O Ensemble implements the Super Learner ensemble (stacking) algorithm using the H2O R interface to provide base learning algorithms. 
}
\details{
\tabular{ll}{
Package: \tab h2oEnsemble\cr
Type: \tab Package\cr
Version: \tab 0.1.0.9000\cr
Date: \tab 2015-10-29\cr
License: \tab Apache License (== 2.0)\cr
}
}
\author{
Erin LeDell

Maintainer: Erin LeDell <erin@h2o.ai>
}
\references{
LeDell, E. (2015) Scalable Ensemble Learning and Computationally Efficient Variance Estimation (Doctoral Dissertation).  University of California, Berkeley, USA.\cr
\url{http://www.stat.berkeley.edu/~ledell/papers/ledell-phd-thesis.pdf}\cr
\cr
van der Laan, M. J., Polley, E. C. and Hubbard, A. E. (2007) Super Learner, Statistical Applications of Genetics and Molecular Biology, 6, article 25. \cr
\url{http://dx.doi.org/10.2202/1544-6115.1309}\cr
\url{http://biostats.bepress.com/ucbbiostat/paper222}\cr
\cr
Breiman, L. (1996) Stacked Regressions, Machine Learning, 24:49-64.\cr
\url{http://dx.doi.org/10.1007/BF00117832}\cr
\url{http://statistics.berkeley.edu/sites/default/files/tech-reports/367.pdf}
}

\keyword{ models }
\seealso{
\code{\link[SuperLearner:SuperLearner]{SuperLearner}}, \code{\link[subsemble:subsemble]{subsemble}}
}
